requirements.txt
llama-index==0.11.10
llama-index-llms-groq==0.2.0
llama-index-embeddings-openai==0.2.5
llama-index-vector-stores-pinecone==0.2.1
pinecone-client==5.0.1
python-dotenv==1.0.0
moviepy==1.0.3
assemblyai==0.33.0
groq==0.11.0
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
